forza_result/requirements.txt
opencv-python
opencv-contrib-python
opencv-python-headless
opencv-contrib-python-headless
flask
pillow
pytesseract